Modernized from DeBot V3
These are the same commands from DeBot V3 that you can find on my github, but just modernized to be compatible with slash commands, and instead of having the gifs in the same file as an array, they are just in a json, which makes it easier to manage. There are some roleplay commands that are still with DeBot V3 that need to be modernized, but I didn't notice until the time of writing.