function-ized getting balance
Instead of making only the command get the balance, a function now is going to be called when another script, or even the balance script is trying to get the balance. This will save headaches later when pay.js is a thing, and not having to copy and paste code, just to get the balance; we can just use a function. Surprisingly, this little change, costed me 25 minutes of my life I will never get back, all because I forgot a comma, and I do not know why I do programming anymore, programming makes me have too many headaches, sos.